Drafts: Fix instance where body of the message is lost.
WorkingMessage.loadDraft() can return a new WorkingMessage object that does not
have the logical mText. This happens when we have gone to select contacts to send
A message to and have loaded the previous draft upon our return in instances where the
user filled out a text message body before filling out any recipients.

Also, prevent the unnecessary settext when the length of the text is 0.

Ticket: QRDL-957
Change-Id: I718a195d43c71a89db2999ef8686be041564546d