QuickMessage: Make useful on lockscreen
Previously, the 'wake and unlock' feature would do
horrible things like disabling the keyguard and never
enabling it. This caused a bunch of different issues.

The Activity would also wake the screen with a wakelock that
never ended. This wakes the phone using the windowmanager flag.

Ticket(s): QRDL-1022 & QRDL-1000
Change-Id: Id53e70f38a3af0fab74016a29abbe7c7693de9c1